+++ home.admin/00mainMenu.sh
@@ -171,13 +171,15 @@ fi
 if [ "${labelbase}" == "on" ]; then
   OPTIONS+=(LABELBASE "Labelbase (UTXO labeling)")
 fi
+if [ "${publicpool}" == "on" ]; then
+  OPTIONS+=(PUBLICPOOL "Public Pool (Bitcoin Solo Mining)")
+fi
 if [ "${tailscale}" == "on" ]; then
   OPTIONS+=(TAILSCALE "Tailscale VPN")
 fi
 if [ "${telegraf}" == "on" ]; then
   OPTIONS+=(TELEGRAF "Telegraf InfluxDB/Grafana Metrics")
 fi
-
 
 # dont offer to switch to "testnet view for now" - so no wswitch back to mainnet needed
 #if [ ${chain} != "main" ]; then
@@ -340,6 +342,9 @@ case $CHOICE in
             ;;
         LABELBASE)
             sudo /home/admin/config.scripts/bonus.labelbase.sh menu
+            ;;
+        PUBLICPOOL)
+            /home/admin/config.scripts/bonus.publicpool.sh menu
             ;;
         TAILSCALE)
             sudo /home/admin/config.scripts/internet.tailscale.sh menu